refact: melhorando manutencao
