+++ src/utils/horarioUtils.js
@@ -9,6 +9,14 @@ export const DIAS = [
   { value: 'sabado',  label: 'Sábado' },
 ];
 
+export const DAY_NAMES = {
+  segunda: 'Segunda-feira',
+  terca: 'Terça-feira',
+  quarta: 'Quarta-feira',
+  quinta: 'Quinta-feira',
+  sexta: 'Sexta-feira',
+}
+
 export const EMPTY_FORM = {
   subject: '', teacher: '', day: 'segunda', period: 'M',
   startSlot: '', endSlot: '', location: '',
@@ -24,7 +32,7 @@ export const getSlotsFromTimes = (startTime, endTime) => {
 };
 
 
-
+//======================= DAILY
 export const DAYS_MAP = ['domingo', 'segunda', 'terca', 'quarta', 'quinta', 'sexta', 'sabado'];
 
 export const formatLongDate = (date) => {
@@ -40,4 +48,51 @@ export const filterClassesByDay = (schedules, date) => {
   return schedules
     .filter(s => s.day?.toLowerCase() === dayName)
     .sort((a, b) => (a.startTime || '00:00').localeCompare(b.startTime || '00:00'));
+};
+
+//===================================== WEEKLY
+export const organizeSchedulesByDay = (schedules) => {
+  const organized = { segunda: [], terca: [], quarta: [],  quinta: [], sexta: [], sabado: [] };
+  
+  if (!schedules) return organized;
+
+  schedules.forEach(schedule => {
+    const day = schedule.day?.toLowerCase();
+    if (organized[day]) organized[day].push(schedule);
+  });
+
+  Object.keys(organized).forEach(day => {
+    organized[day].sort((a, b) => (a.startTime || '00:00').localeCompare(b.startTime || '00:00'));
+  });
+
+  return organized;
+};
+
+export const getVisibleSlots = (TIME_SLOTS, weekSchedule) => {
+  const allSlots = [];
+  ['M', 'T', 'N'].forEach(shift => {
+    const shiftSlots = TIME_SLOTS[shift];
+    if (shiftSlots) {
+      Object.keys(shiftSlots).forEach(key => {
+        allSlots.push({ ...shiftSlots[key], label: `${shift}${key}` });
+      });
+    }
+  });
+
+
+  const findClass = (dayKey, slot) => {
+    const dayClasses = weekSchedule[dayKey];
+    return dayClasses?.find(cls => cls.startTime <= slot.start && cls.endTime >= slot.end);
+  };
+
+  
+  const lastIndex = allSlots.reduce((last, slot, index) => {
+    const hasClass = Object.keys(DAY_NAMES).some(dayKey => !!findClass(dayKey, slot));
+    return hasClass ? index : last;
+  }, -1);
+
+  return {
+    visibleSlots: lastIndex !== -1 ? allSlots.slice(0, lastIndex + 1) : allSlots,
+    findClass
+  };
 };
